# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FlashCash Log API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: List all clients
      parameters:
        - name: transportadoraId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
...
  /viagens/{id}/canhoto:
    post:
      operationId: submitCanhoto
      tags: [canhotos]
      summary: Submit delivery receipt (canhoto) for a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCanhotoInput"
      responses:
        "200":
          description: Canhoto submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canhoto"
  /canhotos:
    get:
      operationId: listCanhotos
      tags: [canhotos]
      summary: List all delivery receipts
      parameters:
        - name: transportadoraId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, validado, rejeitado]
      responses:
        "200":
          description: List of canhotos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Canhoto"
  /canhotos/{id}:
    get:
      operationId: getCanhoto
      tags: [canhotos]
      summary: Get a canhoto by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Canhoto details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canhoto"
  /canhotos/{id}/validate:
    post:
      operationId: validateCanhoto
      tags: [canhotos]
      summary: Manually validate or reject a canhoto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCanhotoInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canhoto"
...
components:
  schemas:
...
    Cliente:
      type: object
      properties:
        id:
          type: integer
        transportadoraId:
          type: integer
        nome:
          type: string
        cnpj:
          type: string
        email:
          type: string
        emailFinanceiro:
          type: string
        telefone:
          type: string
        endereco:
          type: string
        totalFaturas:
          type: integer
        valorTotal:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - transportadoraId
        - nome
        - cnpj
        - email
    CreateClienteInput:
      type: object
      properties:
        transportadoraId:
          type: integer
        nome:
          type: string
        cnpj:
          type: string
        email:
          type: string
        emailFinanceiro:
          type: string
        telefone:
          type: string
        endereco:
          type: string
      required:
        - transportadoraId
        - nome
        - cnpj
        - email
...
    Canhoto:
      type: object
      properties:
        id:
          type: integer
        viagemId:
          type: integer
        motoristaId:
          type: integer
        fotoUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        timestamp:
          type: string
          format: date-time
        numeroCte:
          type: string
        cnpjCliente:
          type: string
        numeroNF:
          type: string
        valorDetectado:
          type: number
        assinaturaDetectada:
          type: boolean
        sealId:
          type: string
        status:
          type: string
          enum: [pendente, validado, rejeitado]
        iaConfidencia:
          type: number
        observacoes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - viagemId
        - motoristaId
        - status
        - timestamp
    SubmitCanhotoInput:
      type: object
      properties:
        fotoUrl:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        numeroCte:
          type: string
        cnpjCliente:
          type: string
        numeroNF:
          type: string
        valorDetectado:
          type: number
        assinaturaDetectada:
          type: boolean
      required:
        - fotoUrl
        - latitude
        - longitude
    ValidateCanhotoInput:
      type: object
      properties:
        status:
          type: string
          enum: [validado, rejeitado]
        observacoes:
          type: string
      required:
        - status
...